ai-service/requirements.txt
# Core Data Processing
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0

# Deep Learning (Task 2: RNN/LSTM/BiLSTM)
tensorflow-cpu==2.13.0
keras==2.13.1

# Graph & Knowledge Base (Task 3: Neo4j)
neo4j==5.14.0
networkx==3.1

# RAG & Language Model (Task 4: GraphRAG)
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
openai==1.3.0

# Vector Database (optional backup)
chromadb==0.4.10
--extra-index-url https://download.pytorch.org/whl/cpu
torch
sentence-transformers==2.2.2

# Web API (Task 5: FastAPI)
fastapi==0.103.0
uvicorn==0.23.2
pydantic==2.3.0

# Visualization (Task 2: Plots)
matplotlib==3.7.2
seaborn==0.12.2

# Utils
python-dotenv==1.0.0
requests==2.31.0

